+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next version
+## 3.5.1
 
 - Add title and date to comments
 - Truncate comments